type refactor
